requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
gitpython==3.1.40
pytest==7.4.3
pytest-asyncio==0.21.1

# Analysis tools
pylint==3.0.2
radon==6.0.1
bandit==1.7.5

# AI/ML detection
tensorflow-hub==0.15.0
torch>=2.2.0
transformers==4.35.2

# Solana specific
solana==0.30.2
anchorpy==0.18.0
